refactor: extract base map config
