--- src/features/geomap/map-styles.ts
@@ -1,1 +1,0 @@
-export const base = 'https://basemaps.cartocdn.com/gl/positron-gl-style/style.json';
