+++ src/features/geomap/maps.config.ts
@@ -1,0 +1,8 @@
+export const base = {
+  mapStyle: 'https://basemaps.cartocdn.com/gl/positron-gl-style/style.json',
+  initialViewState: {
+    longitude: 7.571606,
+    latitude: 50.226913,
+    zoom: 4,
+  },
+};
