+++ src/pages/coordination.page.tsx
@@ -20,6 +20,7 @@ import type { Person } from '@/features/common/entity.model';
 import type { EventType } from '@/features/common/event-types';
 import { eventTypes } from '@/features/common/event-types';
 import { GeoMap } from '@/features/geomap/geo-map';
+import { base as baseMap } from '@/features/geomap/maps.config';
 import { PersonEventsLayer } from '@/features/geomap/person-events-layer';
 import { PinLayer } from '@/features/geomap/PinLayer';
 import { selectZoomToTimeRange } from '@/features/timeline/timeline.slice';
@@ -102,7 +103,7 @@ export default function CoordinationPage(): JSX.Element | null {
           />
         </Grid>
         <Grid item xs={5}>
-          <GeoMap>
+          <GeoMap {...baseMap}>
             {showEventTypes.length >= 2 ? (
               <PersonEventsLayer
                 persons={filteredPersonArray}
